pyproject.toml
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]



[project]
name = "code_agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.9.0",
    "datasets>=3.1.0",
    "faiss-cpu>=1.8.0",
    "fastapi>=0.116.1",
    "huggingface-hub>=0.26.2",
    "ipykernel>=6.29.5",
    "ipython>=8.29.0",
    "jupyter-client>=8.6.3",
    "jupyter-core>=5.7.2",
    "langchain-community>=0.3.21",
    "langchain-huggingface>=0.1.2",
    "langchain-litellm>=0.2.2",
    "langchain-milvus>=0.2.1",
    "langgraph>=0.5.3",
    "langgraph-cli[inmem]>=0.3.5",
    "libcst>=1.5.0",
    "litellm==1.73.0",
    "llama-cloud>=0.1.4",
    "llama-index>=0.11.22",
    "llama-index-agent-openai>=0.3.4",
    "llama-index-cli>=0.3.1",
    "llama-index-core>=0.11.22",
    "llama-index-embeddings-azure-openai>=0.2.5",
    "llama-index-embeddings-openai>=0.2.5",
    "llama-index-indices-managed-llama-cloud>=0.4.0",
    "llama-index-legacy>=0.9.48.post4",
    "llama-index-llms-azure-openai>=0.2.2",
    "llama-index-llms-openai>=0.2.16",
    "llama-index-multi-modal-llms-openai>=0.2.3",
    "llama-index-program-openai>=0.2.0",
    "llama-index-question-gen-openai>=0.2.0",
    "llama-index-readers-file>=0.2.2",
    "llama-index-readers-llama-parse>=0.3.0",
    "llama-index-retrievers-bm25>=0.4.0",
    "llama-parse>=0.5.13",
    "matplotlib>=3.9.2",
    "matplotlib-inline>=0.1.7",
    "modelscope>=1.29.2",
    "neomodel>=5.5.2",
    "numpy>=1.26.4",
    "openai>=1.54.3",
    "pickleshare>=0.7.5",
    "pip>=24.3.1",
    "portalocker>=2.10.1",
    "rapidfuzz>=3.10.1",
    "requests>=2.32.3",
    "scipy>=1.14.1",
    "sentence-transformers>=5.0.0",
    "setuptools>=75.3.0",
    "tokenizers>=0.20.3",
    "toml>=0.10.2",
    "torch>=2.6.0",
    "tornado>=6.4.1",
    "tqdm>=4.67.0",
    "traitlets>=5.14.3",
    "transformers>=4.53.3",
    "tree-sitter==0.21.3",
    "tree-sitter-languages>=1.10.2",
    "wheel>=0.44.0",
    "flask>=3.0.3",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name="pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
